requirements.txt
torch
torchvision
Pillow
tqdm
numpy
opencv-python
scipy